Too many changes
Added Team Games docx to document the whole intention of the game. I've left changes visible so we can see what I have simplified.

Added a test to display the cost of a BuildServer. I can't make this test pass because I can reference the payload.cost outside of place where it is assigned. I suspect this will cause a refactoring.

Added a todo list with the number of major tasks I will attempt to prove next.